Fix compiler warnings.
Switch of compiler warnings for part of the code.
The compiler flag -Weffc++ of the gcc compiler creates many warnings about non virtual destructor of base classes where the
base class is actually a class from boost or from stl. Switch of the compiler flag for the problematic parts of the code
using preprocessor statements. There are also such preprocessor guards for code which creates many warnings when
using clang.

Filter warnings comming from generated files.
When using CTest all warnings comming from generated code (Root Dictionaries, code generated by protoc) will be filtered
before sending the results to the CDash web server.

Remove unused variables or use them.
Initialize all data members in initializer lists.
Use in initializer list the same order of data members as defined in the class declaration.
Declare private copy constructors and assignment operators where needed.

Fix format problems in printf statements.
Correctly cast the variables.

+++ MbsAPI/fLmd.c
@@ -1039,8 +1039,9 @@ uint32_t fLmdOffsetWrite(sLmdControl* pLmdControl)
   }
   if(current/4 != pLmdControl->pMbsFileHeader->iTableOffset) {
     printf("Table offset mismatch: current:%lld calculated:%lld, cur-cal %lld\n",
-           current/4,pLmdControl->pMbsFileHeader->iTableOffset,
-           current/4-pLmdControl->pMbsFileHeader->iTableOffset);
+           (long long int)(current/4),
+           (long long int)(pLmdControl->pMbsFileHeader->iTableOffset),
+           (long long int)(current/4-pLmdControl->pMbsFileHeader->iTableOffset));
     return(LMD__FAILURE);
   }
   if(iReturn != (pLmdControl->iElements+1)*pLmdControl->iOffsetSize) {
@@ -1148,7 +1149,7 @@ void fLmdPrintFileHeader(uint32_t iVerbose, sMbsFileHeader* pMbsFileHeader)
              pMbsFileHeader->iTimeSpecSec,
              pMbsFileHeader->iTimeSpecNanoSec/1000,
              pMbsFileHeader->iMaxWords,
-             pMbsFileHeader->iTableOffset,
+             (long long unsigned int)(pMbsFileHeader->iTableOffset),
              pMbsFileHeader->iOffsetSize);
     }
   }
@@ -1185,7 +1186,7 @@ void fLmdPrintControl(uint32_t iVerbose, sLmdControl* pLmdControl)
            pLmdControl->cFile,
            pLmdControl->iBufferWords,
            pLmdControl->iLeftWords,
-           pLmdControl->iBytes,
+           (long long int)(pLmdControl->iBytes),
            pLmdControl->iElements
           );
     fLmdPrintFileHeader(iVerbose,pLmdControl->pMbsFileHeader);
